[3787] Move classes to utils package
