+++ stream-chat-android-ui-utils/src/main/kotlin/io/getstream/chat/android/uiutils/utils/ColorUtils.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.getstream.chat.android.uiutils
+package io.getstream.chat.android.uiutils.utils
 
 import android.graphics.Color
 import kotlin.math.roundToInt
